⚰️ removendo arquivos mortos
