--- src/pages/Cadastro/formularioTermosDeUso.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { View, Text } from 'react-native';
-
-export default function TermosDeUso() {
-  return (
-    <View>
-      <Text>Termos de uso do ID Saúde</Text>
-    </View>
-  );
-}
