+++ src/pages/NovoCadastro/FormularioInfoProfissional/styles.js
@@ -1,40 +1,28 @@
-import styled from 'styled-components/native';
-import { Title, TextInput, Button, List, Text } from 'react-native-paper';
 import { ScrollView } from 'react-native';
-import IconDropdown from 'react-native-vector-icons/MaterialIcons';
-import featureAtivas from '../../featureAtivas';
-import features from '../../constantes/features';
+import { Button, Title } from 'react-native-paper';
+import styled from 'styled-components/native';
+import features from '~/constantes/features';
+import featureAtivas from '~/featureAtivas';
 
-// TODO: possível remoção
-export const Titulo = styled(Title)`
-  font-size: 24px;
-  line-height: 28px;
-  color: rgba(0, 0, 0, 0.87);
-  margin-top: 16px;
-`;
-
-export const Scroll = styled(ScrollView)`
+export const Container = styled(ScrollView)`
   background-color: #ffffff;
   padding-left: 16px;
   padding-right: 16px;
 `;
 
-export const TituloDoFormulario = styled(Title)`
-  margin-top: 24px;
+export const Titulo = styled(Title)`
+  font-size: 18px;
+  line-height: 28px;
   color: rgba(0, 0, 0, 0.87);
-  font-size: 20px;
+  margin-top: 16px;
+`;
+
+export const SubTitulo = styled(Title)`
+  font-size: 16px;
+  margin-top: 12px;
+  color: rgba(0, 0, 0, 0.87);
   letter-spacing: 0.15px;
   line-height: 23px;
-`;
-
-export const CampoDeTexto = styled(TextInput)`
-  padding-top: 16px;
-  background-color: #fff;
-`;
-
-export const TextoDeErro = styled.Text`
-  color: red;
-  margin-top: 0;
 `;
 
 export const Botao = styled(Button)`
@@ -42,7 +30,7 @@ export const Botao = styled(Button)`
   width: 150px;
   height: 45px;
   align-self: flex-end;
-  margin: 20px;
+  margin-top: 20px;
   justify-content: center;
   background-color: ${props => {
     if (props.disabled) {
@@ -53,26 +41,4 @@ export const Botao = styled(Button)`
     }
     return '#304FFE';
   }};
-`;
-
-export const ConteudoDropdown = styled.View`
-  margin-top: 14px;
-`;
-
-export const IconeDropdown = styled(IconDropdown)`
-  position: absolute;
-  right: 8px;
-  top: 30px;
-  font-size: 25px;
-`;
-
-export const PlaceholderAcordeon = styled(Text)`
-  font-size: 16px;
-  color: rgba(0, 0, 0, 0.54);
-`;
-
-export const Acordeon = styled(List.Accordion)`
-  border-color: rgba(25, 25, 25, 0.32);
-  border-width: 2px;
-  margin-top: 16px;
 `;
