+++ src/theme/index.js
@@ -5,7 +5,6 @@ const theme = {
   colors: {
     ...DefaultTheme.colors,
     primary: '#4CAF50',
-    // background: '#FFF'
   },
 };
